+++ changelog.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `feature` View of cards can be filtered by status.
 - `feature` Setting of favorite entities.
 - `feature` Improve audit items.
+- `feature` Add social audit in event-config view.
 
 ### Fix